requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
beartype==0.14.1
black==21.4b2
bleach==3.0.2
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
cachetools==5.3.2
certifi==2020.6.20
chardet==3.0.4
click==8.1.7
cloudpickle==2.2.1
colored==1.4.4
contextlib2==21.6.0
cvxpy==1.1.5
cycler==0.10.0
dask==0.20.0
decorator==4.3.0
defusedxml==0.5.0
detectron2==0.5+cu110
dill==0.3.7
ecos==2.0.7.post1
entrypoints==0.2.3
flatbuffers==23.5.26
fonttools==4.25.0
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==2.26.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.62.0
graphviz==0.10.1
grpcio==1.60.0
h5py==3.8.0
humanfriendly==10.0
humanize==4.6.0
hydra-core==1.3.2
idna==2.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.8
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
joblib==0.16.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kaggle==1.6.3
keras==2.11.0
kiwisolver==1.0.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib @ file:///croot/matplotlib-suite_1667356714455/work
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
ml-collections==0.1.1
mock==5.1.0
more-itertools==4.3.0
munkres==1.1.4
mypy-extensions==1.0.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numexpr @ file:///croot/numexpr_1668713893690/work
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
osqp==0.6.1
packaging==23.2
pandas==1.3.5
pandocfilters==1.4.2
parso==0.3.1
pathspec==0.11.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.5.0
pluggy==0.8.0
ply==3.11
portalocker==2.7.0
prometheus-client==0.4.2
promise==2.3
prompt-toolkit==2.0.7
protobuf==3.19.6
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydot==2.0.0
Pygments==2.2.0
pyparsing==3.1.1
PyQt5-sip==12.11.0
pyro-api==0.1.2
pyro-ppl==1.8.6
pytest==3.9.3
python-dateutil==2.7.5
python-slugify==8.0.1
pytz==2018.7
PyWavelets==1.0.1
PyYAML==6.0.1
pyzmq==17.1.2
qtconsole==4.4.2
regex==2023.12.25
requests==2.24.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.14.1
scikit-learn==0.23.2
scipy==1.1.0
scs==2.1.2
seaborn @ file:///croot/seaborn_1673479180098/work
Send2Trash==1.5.0
simplejson @ file:///tmp/build/80754af9/simplejson_1639127190616/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
slurm-gpustat==0.0.15
submitit==1.4.6
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.12.0
termcolor==2.3.0
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.3
tfds-nightly==4.4.0.dev202201080107
threadpoolctl==2.1.0
timm==0.4.12
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz==0.9.0
torch==1.13.1
torchaudio==0.7.0a0+a853dff
torchvision @ file:///tmp/build/80754af9/torchvision_1622219711511/work
tornado==5.1.1
tqdm==4.66.1
traitlets==4.3.2
typed-ast==1.5.5
typeguard==4.1.2
typing_extensions==4.7.1
urllib3==1.25.11
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==3.4.2
wrapt==1.16.0
yacs==0.1.8
zipp==3.15.0
